refactoring and layout changes
