+++ src/css/ReactWeather.scss
@@ -1,0 +1,6 @@
+@import "variables";
+.rw-box{
+  font-family: $rw-box-font;
+  font-size:13px;
+}
+
